machine: support arduino-nano33 on play.tinygo.org
