+++ src/machine/board_atsamd21.go
@@ -1,4 +1,4 @@
-// +build sam,atsamd21 circuitplay_express
+// +build sam,atsamd21 arduino_nano33 circuitplay_express
 
 package machine
